feat: stories in sfc
