+++ src/components/atoms/UiTextarea/UiTextarea.stories.ts
@@ -5,20 +5,22 @@ import type {
 import {
   ref,
   provide,
-  inject
+  inject,
 } from 'vue';
-import deepmerge from "deepmerge";
+import deepmerge from 'deepmerge';
 import { UiTextarea } from '@/../index';
-import raw from './UiTextarea.vue?raw';
 import { withVariants } from '@sb/decorators';
 import {
   useArgTypes,
-  inputEvents
-} from '@sb/helpers'
+  inputEvents,
+} from '@sb/helpers';
+
 const { argTypes } = useArgTypes(deepmerge(
   UiTextarea,
-  inputEvents
+  inputEvents,
 ));
+import BasicStories from './stories/Basic.vue';
+import BasicStoriesCode from './stories/Basic.vue?raw';
 
 const meta = {
   title: 'Atoms/Textarea',
@@ -41,146 +43,90 @@ const meta = {
         'horizontal',
         'vertical',
       ],
-    }
+    },
   },
   parameters: {
     chromatic: { disableSnapshot: false },
-    docs: {
-      source: {
-        code: null
-      }
-    }
-  }
+    docs: { source: { code: null } },
+  },
 } satisfies Meta<typeof UiTextarea>;
 export default meta;
 type Story = StoryObj<typeof UiTextarea>;
 
-export const Basic: Story = {
-  render: () => ({
-    inheritAttrs: false,
-    components: { UiTextarea },
-    setup( props, { attrs })  {
-      const {
-        modifiers,
-        modelValue,
-        ...args
-      } = attrs;
-      const value = inject('value') || ref(modelValue);
-
-      return {
-        args: {
-          ...args,
-          class: modifiers,
-        },
-        value,
-      }
-    },
-    template: `<UiTextarea
-      v-model="value"
-      v-bind="args"
-    />`,
-  }),
-};
+export const Basic: Story = { render: () => BasicStories };
 Basic.parameters = {
   chromatic: { disableSnapshot: true },
-  docs: {
-    source: {
-      code: `<template>
-  <UiTextarea
-    v-model="modelValue"
-    :resize="resize"
-    :placeholder="placeholder"
-    :disabled="disabled"
-    textarea-attrs="textareaAttrs"
-  />
-</template>
-
-<script setup lang="ts">
-import { ref } from 'vue';
-import { UiTextarea } from '@infermedica/component-library';
-
-const modelValue = ref('');
-const resize = false;
-const placeholder = 'Please provide a detailed description of the issue.';
-const disabled = false;
-const textareaAttrs = {
-  'data-testid': 'textarea-element'
-}
-</script>`
-    }
-  }
+  docs: { source: { code: BasicStoriesCode } },
 };
 
-export const Empty: Story = {
-  ...Basic
-}
+export const Empty: Story = { ...Basic };
 Empty.argTypes = {
   modelValue: { control: false },
   placeholder: { control: 'text' },
-  disabled: { control: false }
-}
+  disabled: { control: false },
+};
 Empty.decorators = [
   withVariants,
   () => ({
-    setup( props, { attrs }) {
-      const { modelValue } = attrs;
+    setup(props, { attrs }) {
+      const {
+        modelValue, ...args
+      } = attrs;
       const value = ref(modelValue);
       provide('value', value);
     },
-    template: `<story />`
-  })
-]
+    template: '<story />',
+  }),
+];
 Empty.parameters = {
   variants: [
-    {
-      label: 'default',
-    },
-    ...['hover', 'focus-within'].map((variant) => ({
+    { label: 'default' },
+    ...[
+      'hover',
+      'focus-within',
+    ].map((variant) => ({
       label: `${variant}`,
       class: `pseudo-${variant}`,
     })),
     {
       label: 'disabled',
       disabled: true,
-      class: 'ui-textarea--is-disabled'
-    }
+      class: 'ui-textarea--is-disabled',
+    },
   ],
   chromatic: { disableSnapshot: false },
   docs: { source: { code: null } },
-}
-
+};
 
 export const Filled: Story = { ...Empty };
-Filled.args = {
-  modelValue: 'I encountered an error message while trying to submit a form on your website. The error message read \'500 Internal Server Error\'.'
-};
+Filled.args = { modelValue: 'I encountered an error message while trying to submit a form on your website. The error message read \'500 Internal Server Error\'.' };
 Filled.argTypes = {
   ...Empty.argTypes,
   modelValue: { control: 'text' },
   placeholder: { control: false },
-}
+};
 
 export const WithError: Story = { ...Basic };
 WithError.argTypes = {
   modelValue: { control: false },
   placeholder: { control: false },
 };
-WithError.decorators = [ withVariants ]
+WithError.decorators = [ withVariants ];
 WithError.parameters = {
   ...Empty.parameters,
   variants: [
     {
       label: 'default',
-      class: 'ui-textarea--has-error'
+      class: 'ui-textarea--has-error',
     },
-    ...['hover'].map((variant) => ({
+    ...[ 'hover' ].map((variant) => ({
       label: `${variant}`,
       class: `ui-textarea--has-error pseudo-${variant}`,
     })),
     {
       label: 'filed',
       class: 'ui-textarea--has-error',
-      modelValue: 'I encountered an error message while trying to submit a form on your website. The error message read \'500 Internal Server Error.'
-    }
-  ]
-}
+      modelValue: 'I encountered an error message while trying to submit a form on your website. The error message read \'500 Internal Server Error.',
+    },
+  ],
+};
