+++ src/components/molecules/UiNumberStepper/UiNumberStepper.stories.ts
@@ -5,33 +5,21 @@ import type {
 import { ref } from 'vue';
 import {
   UiText,
-  UiNumberStepper
+  UiNumberStepper,
 } from '@/../index';
 import { useArgTypes } from '@sb/helpers';
 
-const { argTypes } = useArgTypes(UiNumberStepper, {
-  variables: {
-    regexp: /^(\.ui-number-stepper|\.ui-number-stepper__decrement|\.ui-number-stepper__increment)$/
-  }
-});
+const { argTypes } = useArgTypes(UiNumberStepper, { variables: { regexp: /^(\.ui-number-stepper|\.ui-number-stepper__decrement|\.ui-number-stepper__increment)$/ } });
 
 const meta = {
   title: 'Molecules/NumberStepper',
   component: UiNumberStepper,
   args: {},
-  argTypes: {
-    ...argTypes
-  },
+  argTypes: { ...argTypes },
   parameters: {
-    chromatic: {
-      disableSnapshot: false,
-    },
-    docs: {
-      source: {
-        code: null,
-      }
-    }
-  }
+    chromatic: { disableSnapshot: false },
+    docs: { source: { code: null } },
+  },
 } satisfies Meta<typeof UiNumberStepper>;
 export default meta;
 type Story = StoryObj<typeof UiNumberStepper>;
@@ -41,9 +29,9 @@ export const Basic: Story = {
     inheritAttrs: false,
     components: {
       UiNumberStepper,
-      UiText
+      UiText,
     },
-    setup: ( props, { attrs } ) => {
+    setup: (props, { attrs }) => {
       const {
         modelValue,
         ...args
@@ -53,7 +41,7 @@ export const Basic: Story = {
       return {
         args,
         value,
-      }
+      };
     },
     template: `<UiNumberStepper
         v-model="value"
@@ -62,6 +50,6 @@ export const Basic: Story = {
       <template #default="{ value }">
         <UiText>{{ value }}</UiText>
       </template>
-    </UiNumberStepper>`
-  })
-}
+    </UiNumberStepper>`,
+  }),
+};
