+++ src/components/organisms/UiSidePanel/UiSidePanel.stories.ts
@@ -45,6 +45,10 @@ type SidePanelMetaType = Meta<SidePanelArgsType>;
 type SidePanelStoryType = StoryObj<SidePanelArgsType>;
 
 const { argTypes } = useArgTypes(UiSidePanel);
+import ConditionDetailsStories from './stories/ConditionDetails.vue';
+import ConditionDetailsStoriesCode from './stories/ConditionDetails.vue?raw';
+import ReportAnIssueStories from './stories/ReportAnIssue.vue';
+import ReportAnIssueStoriesCode from './stories/ReportAnIssue.vue?raw';
 
 const meta = {
   title: 'Organisms/SidePanel',
@@ -337,578 +341,21 @@ onMounted(() => {
   },
 };
 
-export const ConditionDetails: SidePanelStoryType = {
-  render: () => ({
-    inheritAttrs: false,
-    components: {
-      UiSidePanel,
-      UiProgress,
-      UiText,
-      UiHeading,
-      UiIcon,
-      UiButton,
-    },
-    setup(props, { attrs }) {
-      const {
-        modelValue,
-        ...args
-      } = attrs;
-      const value = inject('value');
-      const isShowMoreOpen = ref(false);
-      const handleShowMoreClick = () => {
-        isShowMoreOpen.value = !isShowMoreOpen.value;
-      };
-
-      return {
-        value,
-        args,
-        isShowMoreOpen,
-        handleShowMoreClick,
-      };
-    },
-    template: `<UiSidePanel
-      v-model="value"
-      v-bind="args"
-      class="side-panel-with-condition-details"
-    >
-      <template #title="{ title }">
-        <div class="side-panel-with-condition-details__byline">
-          <UiProgress :value="100"/>
-          <UiText class="ui-text--body-2-comfortable">
-            Strong evidence
-          </UiText>
-        </div>
-        <UiHeading>
-          {{ title }}
-        </UiHeading>
-      </template>
-      <template #default>
-        <div class="side-panel-with-condition-details-section">
-          <UiHeading class="side-panel-with-condition-details-section__title">
-            About
-          </UiHeading>
-          <UiHeading
-            level="4"
-            class="side-panel-with-condition-details-section__subtitle"
-          >
-            What is acute viral throat infection?
-          </UiHeading>
-          <template v-if="isShowMoreOpen">
-            <UiText class="side-panel-with-condition-details-section__paragraph">
-              Acute viral tonsillopharyngitis known as sore throat is an inflammation of throat and tonsils, caused by viral infection. It causes pain, discomfort, swelling in the throat, which can also be red and dry.
-            </UiText>
-            <UiText class="side-panel-with-condition-details-section__paragraph">
-              Other symptoms of pharyngitis may include painful swallowing, fever, bad breath, mild cough, joint pain, sneezing, headache, muscle aches, tender, swollen lymph nodes in the neck or runny nose.
-            </UiText>
-          </template>
-          <template v-else>
-            <UiText class="side-panel-with-condition-details-section__paragraph">Acute viral tonsillopharyngitis known as sore throat is an inflammation of throat and tonsils, caused by viral infection ...</UiText>
-          </template>
-          <UiButton
-              @click="handleShowMoreClick"
-              class="ui-button--text side-panel-with-condition-details-section__show-more"
-          >
-            <UiIcon 
-              :icon="isShowMoreOpen ? 'chevron-up' : 'chevron-down'" 
-              class="ui-button__icon"
-            />Show {{isShowMoreOpen ? 'less' : 'more'}}
-          </UiButton>
-        </div>
-      </template>
-    </UiSidePanel>`,
-  }),
-};
+export const ConditionDetails: SidePanelStoryType = { render: () => (ConditionDetailsStories) };
 ConditionDetails.args = {
   title: 'Acute viral throat infection',
   subtitle: 'Acute viral pharyngitis',
   labelAttrs: { class: 'side-panel-with-condition-details__label' },
   textSubtitleAttrs: { class: 'ui-text--theme-secondary' },
 };
-ConditionDetails.parameters = {
-  docs: {
-    source: {
-      code: `<template>
-  <UiSidePanel
-    v-model="modelValue"
-    :title="title"
-    :subtitle="subtitle"
-    class="side-panel-with-condition-details"
-  >
-    <template #title="{ title }">
-      <div class="side-panel-with-condition-details__byline">
-        <UiProgress :value="100"/>
-        <UiText class="ui-text--body-2-comfortable">
-          Strong evidence
-        </UiText>
-      </div>
-      <UiHeading>
-        {{ title }}
-      </UiHeading>
-    </template>
-    <template #default>
-      <div class="side-panel-with-condition-details-section">
-        <UiHeading class="side-panel-with-condition-details-section__title">
-          About
-        </UiHeading>
-        <UiHeading
-          level="4"
-          class="side-panel-with-condition-details-section__subtitle"
-        >
-          What is acute viral throat infection?
-        </UiHeading>
-        <template v-if="isShowMoreOpen">
-          <UiText class="side-panel-with-condition-details-section__paragraph">
-            Acute viral tonsillopharyngitis known as sore throat is an inflammation of throat and tonsils, caused by viral infection. It causes pain, discomfort, swelling in the throat, which can also be red and dry.
-          </UiText>
-          <UiText class="side-panel-with-condition-details-section__paragraph">
-            Other symptoms of pharyngitis may include painful swallowing, fever, bad breath, mild cough, joint pain, sneezing, headache, muscle aches, tender, swollen lymph nodes in the neck or runny nose.
-          </UiText>
-        </template>
-        <template v-else>
-          <UiText class="side-panel-with-condition-details-section__paragraph">Acute viral tonsillopharyngitis known as sore throat is an inflammation of throat and tonsils, caused by viral infection ...</UiText>
-        </template>
-        <UiButton
-            @click="handleShowMoreClick"
-            class="ui-button--text side-panel-with-condition-details-section__show-more"
-        >
-          <UiIcon 
-            :icon="isShowMoreOpen ? 'chevron-up' : 'chevron-down'" 
-            class="ui-button__icon"
-          />Show {{ isShowMoreOpen ? 'less' : 'more' }}
-        </UiButton>
-      </div>
-    </template>
-  </UiSidePanel>
-</template>
+ConditionDetails.parameters = { docs: { source: { code: ConditionDetailsStoriesCode } } };
 
-<script setup lang="ts">
-import { ref } from 'vue';
-import { 
-  UiSidePanel,
-  UiProgress,
-  UiText,
-  UiHeading,
-  UiIcon,
-  UiButton,
-} from '@infermedica/component-library';
-import type { SidePanelProps } from '@infermedica/component-library';
-
-const modelValue = ref<SidePanelProps['modelValue']>(${meta.args.modelValue});
-const title: SidePanelProps['title'] = '${ConditionDetails.args.title}';
-const subtitle: SidePanelProps['subtitle'] = '${ConditionDetails.args.subtitle}';
-const labelAttr: SidePanelProps['labelAttrs'] = '${ConditionDetails.args.labelAttrs}';
-const textSubtitleAttrs: SidePanelProps['textSubtitleAttrs'] = '${ConditionDetails.args.textSubtitleAttrs}';
-const isShowMoreOpen = ref(false);
-const handleShowMoreClick = () => {
-  isShowMoreOpen.value = !isShowMoreOpen.value;
-};
-</script>
-
-<style lang="scss">
-@use "@infermedica/component-library/mixins";
-
-.side-panel-with-condition-details {
-  &__label {
-    display: grid;
-    gap: var(--space-4);
-  }
-
-  &__byline {
-    display: grid;
-    grid-template-columns: 4rem 1fr;
-    align-items: center;
-    gap: var(--space-12);
-  }
-
-  &-section {
-    @include mixins.inner-border(
-      'side-panel-with-condition-details-section',
-      $color: var(--color-border-subtle),
-      $width: 0 0 1px 0,
-      $radius: 0
-    );
-
-    &__title {
-      margin-block-end: var(--space-12);
-    }
-
-    &__paragraph {
-      margin-block-start: var(--space-12);
-    }
-
-    &__show-more {
-      margin-block: var(--space-24) var(--space-32);
-    }
-  }
-}
-</style>`,
-    },
-  },
-};
-
-export const ReportAnIssue: SidePanelStoryType = {
-  render: () => ({
-    inheritAttrs: false,
-    components: {
-      UiSidePanel,
-      UiText,
-      UiFormField,
-      UiList,
-      UiLoader,
-      UiCheckbox,
-      UiTextarea,
-      UiButton,
-    },
-    setup(props, { attrs }) {
-      const {
-        modelValue,
-        ...args
-      } = attrs;
-      const value = inject('value');
-      const options = [
-        {
-          name: 'too-long',
-          label: 'It\'s too long',
-        },
-        {
-          name: 'unrelated',
-          label: 'It seems to be unrelated to me',
-        },
-        {
-          name: 'complex',
-          label: 'It contains complex words',
-        },
-        {
-          name: 'answered',
-          label: 'I\'ve already answered this question',
-        },
-        {
-          name: 'which-answer-to-choose',
-          label: 'I can\'t decide which answer to choose',
-        },
-        {
-          name: 'typo',
-          label: 'I found a typo',
-        },
-        {
-          name: 'other',
-          label: 'Other (please comment below)',
-        },
-      ].map((option) => ({
-        ...option,
-        listItemAttrs: { class: 'side-panel-with-report-an-issue__option' },
-      }));
-      const selected = ref([]);
-      const isProcessing = ref(false);
-      const isTouched = ref(false);
-      const errorMessage = computed(() => (
-        selected.value.length > 0 || !isTouched.value
-          ? false
-          : 'Please select at least one issue.'
-      ));
-      const closeHandler = () => {
-        selected.value.length = 0;
-        isTouched.value = false;
-
-        (value as Ref<unknown>).value = false;
-      };
-      const handleReportAnIssueSubmit = () => {
-        isTouched.value = true;
-        if (!errorMessage.value) {
-          isProcessing.value = true;
-          setTimeout(() => {
-            isProcessing.value = false;
-            closeHandler();
-          }, 1000);
-        }
-      };
-
-      return {
-        value,
-        args,
-        options,
-        selected,
-        isProcessing,
-        errorMessage,
-        handleReportAnIssueSubmit,
-        closeHandler,
-      };
-    },
-    template: `<UiSidePanel
-      v-model="value"
-      v-bind="args"
-      class="side-panel-with-report-an-issue"
-    >
-      <form @submit.prevent="handleReportAnIssueSubmit">
-        <UiText>
-          What's wrong with this question?
-        </UiText>
-        <UiFormField 
-          :error-message="errorMessage"
-          :alert-attrs="{
-            class: 'side-panel-with-report-an-issue__alert'
-          }"
-        >
-          <UiList
-            :items="options"
-            class="side-panel-with-report-an-issue__options"
-          >
-            <template
-              v-for="({name}, key) in options"
-              #[name]="{item: { label }}"
-              :key="key"
-            >
-              <UiCheckbox
-                v-model="selected"
-                :value="label"
-                :disabled="isProcessing"
-                :class="{
-                  'ui-checkbox--is-disabled': isProcessing,
-                }"
-              >
-                {{ label }}
-              </UiCheckbox>
-            </template>
-          </UiList>
-        </UiFormField>
-        <UiFormField
-          :errorMessage="false"
-          message="Describe details"
-          hint="Optional"
-          class="side-panel-with-report-an-issue__details"
-        >
-          <template #default="{id}">
-            <UiTextarea
-              :id="id"
-              :disabled="isProcessing"
-            />
-          </template>
-        </UiFormField>
-        <div class="side-panel-with-report-an-issue__actions">
-          <UiButton
-            type="submit"
-            :class="{'ui-button--is-disabled': isProcessing}"
-            :disabled="isProcessing"
-          >
-            <UiLoader
-              :is-loading="isProcessing"
-              type="ellipsis"
-              transition-type="opacity"
-              :transition-attrs="{ name: '' }"
-            >
-              <span>Submit</span>
-            </UiLoader>
-          </UiButton>
-          <UiButton
-            type="button"
-            :class="['ui-button--text', { 'ui-button--is-disabled': isProcessing }]"
-            :disabled="isProcessing"
-            @click="closeHandler"
-          >
-            Cancel
-          </UiButton>
-        </div>
-      </form>
-    </UiSidePanel>`,
-  }),
-};
+export const ReportAnIssue: SidePanelStoryType = { render: () => (ReportAnIssueStories) };
 ReportAnIssue.args = {
   title: 'Report an issue',
   subtitle: '',
 };
-ReportAnIssue.parameters = {
-  docs: {
-    source: {
-      code: `<template>
-  <UiSidePanel
-    v-model="modelValue"
-    :title="title"
-    class="side-panel-with-report-an-issue"
-  >
-    <form @submit.prevent="handleReportAnIssueSubmit">
-      <UiText>
-        What's wrong with this question?
-      </UiText>
-      <UiFormField 
-        :error-message="errorMessage"
-        :alert-attrs="{
-          class: 'side-panel-with-report-an-issue__alert'
-        }"
-      >
-        <UiList
-          :items="options"
-          class="side-panel-with-report-an-issue__options"
-        >
-          <template
-            v-for="({name}, key) in options"
-            #[name]="{item: { label }}"
-            :key="key"
-          >
-            <UiCheckbox
-              v-model="selected"
-              :value="label"
-              :disabled="isProcessing"
-              :class="{
-                'ui-checkbox--is-disabled': isProcessing,
-              }"
-            >
-              {{ label }}
-            </UiCheckbox>
-          </template>
-        </UiList>
-      </UiFormField>
-      <UiFormField
-        :errorMessage="false"
-        message="Describe details"
-        hint="Optional"
-        class="side-panel-with-report-an-issue__details"
-      >
-        <template #default="{ id }">
-          <UiTextarea
-            :id="id"
-            :disabled="isProcessing"
-          />
-        </template>
-      </UiFormField>
-      <div class="side-panel-with-report-an-issue__actions">
-        <UiButton
-          type="submit"
-          :class="{'ui-button--is-disabled': isProcessing}"
-          :disabled="isProcessing"
-        >
-          <UiLoader
-            :is-loading="isProcessing"
-            type="ellipsis"
-            transition-type="opacity"
-            :transition-attrs="{ name: '' }"
-          >
-            <span>Submit</span>
-          </UiLoader>
-        </UiButton>
-        <UiButton
-          type="button"
-          :class="['ui-button--text', { 'ui-button--is-disabled': isProcessing }]"
-          :disabled="isProcessing"
-        >
-          Cancel
-        </UiButton>
-      </div>
-    </form>
-  </UiSidePanel>
-</template>
-
-<script setup lang="ts">
-import { 
-  ref,
-  computed,
-} from 'vue';
-import { 
-  UiSidePanel,
-  UiText,
-  UiFormField,
-  UiList,
-  UiCheckbox,
-  UiTextarea,
-  UiButton,
-} from '@infermedica/component-library';
-import type { 
-  SidePanelProps,
-  ListItem,
-} from '@infermedica/component-library';
-
-const modelValue = ref<SidePanelProps['modelValue']>(${meta.args.modelValue});
-const title: SidePanelProps['title'] = '${ReportAnIssue.args.title}';
-const options = ListItem['items'] = [
-    {
-      name: 'too-long',
-      label: 'It's too long',
-    },
-    {
-      name: 'unrelated',
-      label: 'It seems to be unrelated to me',
-    },
-    {
-      name: 'complex',
-      label: 'It contains complex words',
-    },
-    {
-      name: 'answered',
-      label: 'I've already answered this question',
-    },
-    {
-      name: 'which-answer-to-choose',
-      label: 'I can't decide which answer to choose',
-    },
-    {
-      name: 'typo',
-      label: 'I found a typo',
-    },
-    {
-      name: 'other',
-      label: 'Other (please comment below)',
-    },
-  ].map((option) => ({
-    ...option,
-    listItemAttrs: { class: 'side-panel-with-report-an-issue__option' },
-  }));
-const selected = ref([]);
-const isProcessing = ref(false);
-const isTouched = ref(false);
-const errorMessage = computed(() => (
-  selected.value.length > 0 
-  || !isTouched.value 
-    ? false 
-    : 'Please select at least one issue.'
-));
-const handleReportAnIssueSubmit = () => {
-  isTouched.value = true;
-  if (!errorMessage.value) {
-    isProcessing.value = true;
-    // TODO: remove in production code / BEGIN
-    setTimeout(() => {
-      isProcessing.value = false;
-    }, 1000);
-    // END
-  }
-};
-</script>
-
-<style lang="scss">
-.side-panel-with-report-an-issue {
-  &__options {
-    display: flex;
-    flex-direction: column;
-    gap: var(--space-24);
-    margin-block-start: var(--space-24)
-  }
-
-  &__option {
-    --list-item-content-padding-block: 0;
-    --list-item-content-padding-inline: 0;
-    --list-item-tablet-content-padding-block: 0;
-    --list-item-tablet-content-padding-inline: 0;
-    --list-item-border-block-width: 0;
-    --list-item-content-hover-background: transparent;
-  }
-
-  &__details {
-    margin-block: var(--space-32);
-  }
-
-  &__actions {
-    display: flex;
-    align-items: center;
-    gap: var(--space-32);
-  }
-
-  &__alert {
-    --form-field-alert-margin-block-start: var(--space-24);
-  }
-}
-</style>`,
-    },
-  },
-};
+ReportAnIssue.parameters = { docs: { source: { code: ReportAnIssueStoriesCode } } };
 
 export const WithBackdropSlot: SidePanelStoryType = {
   render: () => ({
