+++ src/components/organisms/UiTabs/UiTabs.stories.ts
@@ -43,6 +43,10 @@ const items: TabsItemAttrsProps[] = [
 ];
 
 const { argTypes } = useArgTypes(UiTabs);
+import BasicStories from './stories/Basic.vue';
+import BasicStoriesCode from './stories/Basic.vue?raw';
+import WithDefaultSlotStories from './stories/WithDefaultSlot.vue';
+import WithDefaultSlotStoriesCode from './stories/WithDefaultSlot.vue?raw';
 
 const meta = {
   title: 'Organisms/Tabs',
@@ -70,295 +74,14 @@ const meta = {
 
 export default meta;
 
-export const Basic: TabsStoryType = {
-  render: () => ({
-    inheritAttrs: false,
-    components: {
-      UiTabs,
-      UiText,
-    },
-    setup(props, { attrs }) {
-      const {
-        modelValue,
-        content,
-        items,
-        modifiers,
-        ...args
-      } = attrs;
-
-      const current = ref(modelValue);
-
-      return {
-        current,
-        items,
-        content,
-        args: {
-          ...args,
-          class: modifiers,
-        },
-      };
-    },
-    template: `<UiTabs
-    v-bind="args"
-    v-model="current"
-    :items="items"
-    >
-    <template
-      v-for="(item, key) in items"
-      #[item.name]="{ item }"
-      :key="key"
-    >
-      <UiText>
-        {{ content[item.name] }}
-      </UiText>
-    </template>
-  </UiTabs>`,
-  }),
-};
-Basic.parameters = {
-  docs: {
-    source: {
-      code: `<template>
-    <UiTabs
-    v-model="modelValue"
-    :items="items"
-    :class="modifiers"
-    >
-    <template
-      v-for="(item, key) in items"
-      #[item.name]="{item}"
-      :key="key"
-    >
-      <UiText>
-        {{ content[item.name] }}
-      </UiText>
-    </template>
-    </UiTabs>
-  </template>
-  
-  <script setup lang="ts">
-  import { ref } from 'vue';
-  import {
-    UiTabs,
-    UiText,
-    TabsProps,
-  } from '@infermedica/component-library';
-  
-  const modelValue = ref<TabsProps['modelValue']>('point');
-  const items: TabsProps['items'] = [
-    {
-      name: 'search',
-      title: 'Search',
-      'data-testid': 'search',
-      buttonTabAttrs: { 'data-testid': 'search-button' },
-      contentAttrs: { 'data-testid': 'search-content' },
-    },
-    {
-      name: 'point',
-      title: 'Point on the body',
-      buttonTabAttrs: { 'data-testid': 'point-button' },
-      contentAttrs: { 'data-testid': 'point-content' },
-    },
-    {
-      name: 'drop',
-      title: 'Drop body point',
-      buttonTabAttrs: { 'data-testid': 'drop-button' },
-      contentAttrs: { 'data-testid': 'drop-content' },
-    },
-  ];
-  const content = {
-    search: 'Serum uric acid concentration',
-    point: '1. Erythrocyte Sedimentation Rate',
-  };
-  const modifiers = '';
-  </script>
-  `,
-    },
-  },
-};
+export const Basic: TabsStoryType = { render: () => (BasicStories) };
+Basic.parameters = { docs: { source: { code: BasicStoriesCode } } };
 
 export const Fixed: TabsStoryType = { ...Basic };
 Fixed.args = { modifiers: [ 'ui-tabs--fixed' ] };
-Fixed.parameters = {
-  docs: {
-    source: {
-      code: `<template>
-      <UiTabs
-        v-model="modelValue"
-        :items="items"
-        class="ui-tabs--fixed"
-      >
-      <template
-        v-for="(item, key) in items"
-        #[item.name]="{ item }"
-        :key="key"
-      >
-        <UiText>
-          {{ content[item.name] }}
-        </UiText>
-      </template>
-      </UiTabs>
-    </template>
-    
-    <script setup lang="ts">
-    import { ref } from 'vue';
-    import {
-      UiTabs,
-      UiText,
-    } from '@infermedica/component-library';
-    import type { TabsProps } from '@infermedica/component-library';
-    
-    const modelValue = ref<TabsProps['modelValue']>('point');
-    const items: TabsProps['items'] = [
-      {
-        name: 'search',
-        title: 'Search',
-        'data-testid': 'search',
-        buttonTabAttrs: { 'data-testid': 'search-button' },
-        contentAttrs: { 'data-testid': 'search-content' },
-      },
-      {
-        name: 'point',
-        title: 'Point on the body',
-        buttonTabAttrs: { 'data-testid': 'point-button' },
-        contentAttrs: { 'data-testid': 'point-content' },
-      },
-      {
-        name: 'drop',
-        title: 'Drop body point',
-        buttonTabAttrs: { 'data-testid': 'drop-button' },
-        contentAttrs: { 'data-testid': 'drop-content' },
-      },
-    ];
-    const content = {
-      search: 'Serum uric acid concentration',
-      point: '1. Erythrocyte Sedimentation Rate',
-    };
-    </script>
-    `,
-    },
-  },
-};
 
-export const WithDefaultSlot: TabsStoryType = {
-  render: () => ({
-    inheritAttrs: false,
-    components: {
-      UiTabs,
-      UiTabsItem,
-      UiText,
-    },
-    setup: (props, { attrs }) => {
-      const {
-        modelValue,
-        content,
-        items,
-        modifiers,
-        ...args
-      } = attrs;
-
-      const current = ref(modelValue);
-
-      return {
-        current,
-        items,
-        content,
-        args: {
-          ...args,
-          class: modifiers,
-        }
-        ,
-      };
-    },
-    template: `<UiTabs
-    v-model="current"
-    >
-    <template
-      v-for="(item, key) in items"
-      :key="key"
-    >
-      <UiTabsItem
-        :title="item.title"
-        :name="item.name"
-        :button-attrs="item.buttonTabAttrs"
-        :content-attrs="item.contentAttrs"
-        v-bind="{
-          'data-testid': item['data-testid'],
-        }"
-      >
-        <UiText>
-          {{ content[item.name] }}
-        </UiText>
-      </UiTabsItem>
-    </template>
-  </UiTabs>`,
-  }),
-};
+export const WithDefaultSlot: TabsStoryType = { render: () => (WithDefaultSlotStories) };
 WithDefaultSlot.parameters = {
   chromatic: { disableSnapshot: true },
-  docs: {
-    source: {
-      code: `<template>
-    <UiTabs
-      v-model="modelValue"
-      :class="modifiers"
-    >
-    <template
-      v-for="(item, key) in items"
-      :key="key"
-    >
-      <UiTabsItem
-        :title="item.title"
-        :name="item.name"
-        :button-attrs="item.buttonTabAttrs"
-        :content-attrs="item.contentAttrs"
-        :data-testid="item['data-testid']"
-      >
-        <UiText>
-          {{ content[item.name] }}
-        </UiText>
-      </UiTabsItem>
-    </template>
-    </UiTabs>
-  </template>
-  
-  <script setup lang="ts">
-  import { ref } from 'vue';
-  import {
-    UiTabs,
-    UiText,
-  } from '@infermedica/component-library';
-  import type { TabsProps } from '@infermedica/component-library'
-  import UiTabsItem from '@infermedica/component-library/src/components/organisms/UiTabs/_internal/UiTabsItem.vue';
-  
-  const modelValue = ref<TabsProps['modelValue']>('point');
-  const items: TabsProps['items'] = [
-    {
-      name: 'search',
-      title: 'Search',
-      'data-testid': 'search',
-      buttonTabAttrs: { 'data-testid': 'search-button' },
-      contentAttrs: { 'data-testid': 'search-content' },
-    },
-    {
-      name: 'point',
-      title: 'Point on the body',
-      buttonTabAttrs: { 'data-testid': 'point-button' },
-      contentAttrs: { 'data-testid': 'point-content' },
-    },
-    {
-      name: 'drop',
-      title: 'Drop body point',
-      buttonTabAttrs: { 'data-testid': 'drop-button' },
-      contentAttrs: { 'data-testid': 'drop-content' },
-    },
-  ];
-  const content = {
-    search: 'Serum uric acid concentration',
-    point: '1. Erythrocyte Sedimentation Rate',
-  };
-  const modifiers = '';
-  </script>`,
-    },
-  },
+  docs: { source: { code: WithDefaultSlotStoriesCode } },
 };
